requirements_py310.txt
APScheduler==3.6.3
asgiref==3.2.10
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
bokeh==2.2.3
boto3==1.15.6
botocore==1.18.18
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==3.2.1
# defer==1.0.6
distro==1.4.0
Django==3.1.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-chartjs==2.2.1
django-crontab==0.7.1
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-mysql==3.9.0
django-pandas==0.6.2
enc-dotenv==0.0.3
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
httplib2==0.14.0
idna==2.10
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
keyring==18.0.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
monotonic==1.5
mysql-connector-python==8.0.22
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandas-bokeh==0.5
paramiko==2.6.0
pexpect==4.6.0
Pillow==8.0.1
protobuf==3.13.0
pyAesCrypt==0.4.3
pycparser==2.20
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==0.10.1
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pytesseract==0.3.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.24.0
s3transfer==0.3.3
schedule==0.6.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
threadpoolctl==2.1.0
tornado==6.0.4
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
